requirements.txt
SQLAlchemy~=2.0.40
passlib~=1.7.4
python-jose~=3.4.0
python-dotenv~=1.1.0
fastapi~=0.115.12
pydantic~=2.11.3